refactor preprocessing into modular package
